fixed from doc review, and add fixed example for js/biased-cryptographic-random using a secure library
