+++ javascript/ql/src/Security/CWE-327/examples/bad-random-fixed.js
@@ -1,10 +1,3 @@
-const crypto = require('crypto');
+const cryptoRandomString = require('crypto-random-string');
 
-const digits = [];
-while (digits.length < 10) {
-    const byte = crypto.randomBytes(1)[0];
-    if (byte >= 250) {
-        continue;
-    }
-    digits.push(byte % 10); // OK
-}
+const digits = cryptoRandomString({length: 10, type: 'numeric'});
